Prioritize cloud bucket workflow in setup guide.
The removed `dockerd` command is readily available as the first thing you read on the provided link https://docs.docker.com/config/daemon/#start-the-daemon-manually. I think it makes sense to provide this resource to users who may need it, but personally don't believe it needs a place in the "recommended setup" workflow.